requirements.txt
django==5.0
djangorestframework
pillow
django-filter
django-cors-headers
google-auth-oauthlib
google-auth-httplib2 
google-api-python-client